added controllers, guards, interceptors

+++ src/services/user.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, OnModuleInit } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+  OnModuleInit,
+} from '@nestjs/common';
 import { PrismaService } from './prisma.service';
 import { Prisma, Roles, User } from 'generated/prisma/client';
 import { hashPassword } from '@utils/HashPassword';
@@ -43,7 +48,7 @@ export class UserService implements OnModuleInit {
   async findUserById(
     id: string,
   ): Promise<Omit<User, 'password' | 'isActive'> | null> {
-    return this.prisma.user.findUnique({
+    const user = this.prisma.user.findUnique({
       where: { id },
       select: {
         id: true,
@@ -55,11 +60,14 @@ export class UserService implements OnModuleInit {
         role: true,
       },
     });
+    if (!user) throw new NotFoundException('User not found');
+    return user;
   }
+
   async findByEmail(
     email: string,
   ): Promise<Omit<User, 'password' | 'isActive'> | null> {
-    return this.prisma.user.findUnique({
+    const user = await this.prisma.user.findUnique({
       where: { email },
       select: {
         id: true,
@@ -71,10 +79,19 @@ export class UserService implements OnModuleInit {
         role: true,
       },
     });
+    if (!user)
+      throw new NotFoundException('User not found');
+    return user;
   }
 
   async createUser(data: Prisma.UserCreateInput): Promise<User> {
     const email = data.email.trim().toLowerCase();
+
+    const existingUser = await this.prisma.user.findUnique({
+      where: { email: data.email },
+    });
+    if (existingUser) throw new BadRequestException('Email already in use');
+
     const hashedPassword = await hashPassword(data.password);
     return this.prisma.user.create({
       data: {
@@ -86,6 +103,8 @@ export class UserService implements OnModuleInit {
   }
 
   async updateUser(id: string, data: Partial<UserDto>): Promise<User> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) throw new NotFoundException('User not found');
     const { email } = data;
     if (email) {
       data.email = email.trim().toLowerCase();
@@ -97,6 +116,8 @@ export class UserService implements OnModuleInit {
   }
 
   async UpdateUserPassword(id: string, newPassword: string): Promise<User> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) throw new NotFoundException('User not found');
     const hashedPassword = await hashPassword(newPassword);
     return this.prisma.user.update({
       where: { id },
@@ -104,9 +125,15 @@ export class UserService implements OnModuleInit {
     });
   }
 
-  async deleteUser(where: Prisma.UserWhereUniqueInput): Promise<User> {
+  async deleteUser(id: string): Promise<User> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) throw new NotFoundException('User not found');
     return this.prisma.user.delete({
-      where,
+      where: { id },
     });
+  }
+
+  async getUsersRoles(): Promise<string[]> {
+    return Object.values(Roles);
   }
 }
